package/sunxi-tools: add tool selection
This patch allows to select which of the sunxi-tools are built and installed in host and target diretories.
This way it is possible to have e.g. sunxi-nand-image-builder among the host tools, or to install sunxi-fel on the target device.
The corresponding config options have been added to Config.in and Config.in.host and the sunxi-tools.mk has been adapted respectively.
The defaults have been selected to produce the same output as without this patch.

Signed-off-by: Alex Kaplan <kaplan2539@gmail.com>